option menu only has two buttons (return and exit)

+++ src/editor_env_sector_buttons.c
@@ -6,7 +6,7 @@
 /*   By: sipatry <sipatry@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/22 08:54:19 by sipatry           #+#    #+#             */
-/*   Updated: 2020/01/31 17:20:41 by sipatry          ###   ########.fr       */
+/*   Updated: 2020/02/03 18:03:58 by sipatry          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,5 +37,5 @@ void	init_sector_general_env(t_env *env)
 	env->editor.hud.g_sector.t_intensity.type = INT;
 	env->editor.hud.g_sector.t_gravity.env = env;
 	env->editor.hud.g_sector.t_gravity.pos = new_point(0, 0);
-	env->editor.hud.g_sector.t_gravity.type = INT;
+	env->editor.hud.g_sector.t_gravity.type = DOUBLE;
 }
